requirements.txt
flask==2.3.3
python-docx==0.8.11
PyPDF2==2.10.0
nltk==3.6.3
werkzeug==2.3.7
jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
gunicorn==21.2.0 